requirements.txt
python-telegram-bot==21.11.1
matplotlib==3.10.1
Unidecode==1.3.8
APScheduler==3.11.0
nest-asyncio==1.6.0